refactor(storage): introduce repositories
